Add ContributionVoter
