+++ BackofficeBundle/Resources/config/voter.yml
@@ -1,7 +1,8 @@
 parameters:
-    open_orchestra_backoffice.voter.node.class:       OpenOrchestra\Backoffice\Security\Authorization\Voter\NodeVoter
-    open_orchestra_backoffice.voter.content.class:    OpenOrchestra\Backoffice\Security\Authorization\Voter\ContentVoter
-    open_orchestra_backoffice.voter.trash_item.class: OpenOrchestra\Backoffice\Security\Authorization\Voter\TrashItemVoter
+    open_orchestra_backoffice.voter.node.class:          OpenOrchestra\Backoffice\Security\Authorization\Voter\NodeVoter
+    open_orchestra_backoffice.voter.content.class:       OpenOrchestra\Backoffice\Security\Authorization\Voter\ContentVoter
+    open_orchestra_backoffice.voter.trash_item.class:    OpenOrchestra\Backoffice\Security\Authorization\Voter\TrashItemVoter
+    open_orchestra_backoffice.voter.configuration.class: OpenOrchestra\Backoffice\Security\Authorization\Voter\ConfigurationVoter
 
 services:
     open_orchestra_backoffice.voter.node:
@@ -14,5 +15,9 @@ services:
             - { name: security.voter }
     open_orchestra_backoffice.voter.trash_item:
        class: '%open_orchestra_backoffice.voter.trash_item.class%'
+       tags:
+            - { name: security.voter }
+    open_orchestra_backoffice.voter.configuration:
+       class: '%open_orchestra_backoffice.voter.configuration.class%'
        tags:
             - { name: security.voter }
